feat: LineRequest, LineResponse 수정
- LineRequest
  - up/downStationId Long 타입으로 변경
- LineResponse
  - stations List<Station> 타입으로 변경
  - 생성자 수정

+++ src/main/java/subway/line/LineRequest.java
@@ -3,8 +3,8 @@ package subway.line;
 public class LineRequest {
     private String name;
     private String color;
-    private int upStationId;
-    private int downStationId;
+    private Long upStationId;
+    private Long downStationId;
     private int distance;
 
     public String getName() {
@@ -15,11 +15,11 @@ public class LineRequest {
         return color;
     }
 
-    public int getUpStationId() {
+    public Long getUpStationId() {
         return upStationId;
     }
 
-    public int getDownStationId() {
+    public Long getDownStationId() {
         return downStationId;
     }
